restructured module; adapted translate and validate to the new structure
